new slope viz ready
